Refactor log module into separate files
